chore: convert the connect spec to vitest
